app/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
certifi==2026.1.4
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
playwright==1.50.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.1.1
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzlocal==5.3.1
